added markov chain class w header declarations

+++ main.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 
-vector<vector<int>> multiplyMatrix(const vector<vector<int>>& matrix1, const vector<vector<int>>& matrix2 )
+vector<vector<double>> multiplyMatrix(const vector<vector<double>>& matrix1, const vector<vector<double>>& matrix2 )
 {
     const int n1 = matrix1.size();
     const int m1 = matrix1[0].size();
@@ -12,14 +12,14 @@ vector<vector<int>> multiplyMatrix(const vector<vector<int>>& matrix1, const vec
     const int m2 = matrix2[0].size();
     // matrix 2 is nn x mm
 
-    // to multiply the matricies m1 has to be equal to n2
+    // to multiply the matrices m1 has to be equal to n2
     if (m1 != n2)
     {
         cout << "Matrix multiplication not possible" << endl;
         return {};
     }
 
-    vector result(n1, vector<int>(m2, 0));
+    vector result(n1, vector<double>(m2, 0));
 
     for (int i = 0; i < n1; i++) {
         for (int j = 0; j < m2; j++) {
@@ -35,10 +35,9 @@ vector<vector<int>> multiplyMatrix(const vector<vector<int>>& matrix1, const vec
 
 int main()
 {
-
     //test
-    vector<vector<int>> m1 = { {1, 1}, {2, 2} };
-    vector<vector<int>> m2 = { {1, 1}, {2, 2} };
+    vector<vector<double>> m1 = { {1, 1}, {2, 2} };
+    vector<vector<double>> m2 = { {1, 1}, {2, 2} };
 
     auto result = multiplyMatrix(m1, m2);
 
@@ -50,9 +49,6 @@ int main()
         }
         cout << endl;
     }
-
-
-    
     return 0;
 }
